Change file organization and fix rating filter functionality
