+++ src/components/ratings/Reviews/AddReview/Characteristics.ts
