+++ src/components/ratings/Reviews/AddReview/CharacteristicsList.tsx
